+++ src/binary64/exp/exp.c
@@ -90,7 +90,7 @@ static inline double as_ldexp(double x, i64 i){
 
 static inline double as_todenormal(double x){
 #ifdef __x86_64__
-    __m128i sb; sb[0] = ~0ul>>12;
+    __m128i sb; sb[0] = ~(u64)0>>12;
 #if defined(__clang__)
     __m128d r = _mm_set_sd(x);
 #else
@@ -100,7 +100,7 @@ static inline double as_todenormal(double x){
     return r[0];
 #else
     b64u64_u ix = {.f = x};
-    ix.u &= ~0ul>>12;
+    ix.u &= ~(u64)0>>12;
     return ix.f;
 #endif
 }
@@ -248,7 +248,7 @@ static double __attribute__((noinline)) as_exp_accurate(double x){
       fh = fasttwosum(th,fh, &e);
       fl = fasttwosum(e, fl, &e);
       ix.f = fl;
-      if((ix.u&(~0ul>>12))==0) {
+      if((ix.u&(~(u64)0>>12))==0) {
 	b64u64_u v = {.f = e};
 	i64 d = ((((i64)ix.u>>63)^((i64)v.u>>63))<<1) + 1;
 	ix.u += d;
@@ -260,7 +260,7 @@ static double __attribute__((noinline)) as_exp_accurate(double x){
     }
     fh = fasttwosum(fh,fl, &fl);
     ix.f = fl;
-    u64 d = (ix.u + 2)&(~0ul>>12);
+    u64 d = (ix.u + 2)&(~(u64)0>>12);
     if(__builtin_expect(d<=2, 0)) fh = as_exp_database(x, fh);
     fh = as_ldexp(fh, ie);
   }
@@ -269,7 +269,7 @@ static double __attribute__((noinline)) as_exp_accurate(double x){
 
 double cr_exp(double x){
   b64u64_u ix = {.f = x};
-  u64 aix = ix.u & (~0ul>>1);
+  u64 aix = ix.u & (~(u64)0>>1);
   if(__builtin_expect(aix == 0, 0)) return 1.0;
   if(__builtin_expect(aix>=0x40862e42fefa39f0ul, 0)){
     if(aix>0x7ff0000000000000ul) return x;
